connection.c: try_ready_normal_body(): fix VS compiler warning

+++ src/microhttpd/connection.c
@@ -358,7 +358,7 @@ try_ready_normal_body (struct MHD_Connection *connection)
   ret = response->crc (response->crc_cls,
                        connection->response_write_position,
                        response->data,
-                       MHD_MIN (response->data_buffer_size,
+                       (size_t)MHD_MIN ((uint64_t)response->data_buffer_size,
                                 response->total_size -
                                 connection->response_write_position));
   if ( (((ssize_t) MHD_CONTENT_READER_END_OF_STREAM) == ret) ||
